requirements.txt
requests
PyYAML
Jinja2
selenium
browsermob-proxy
rich
textual
